face_service/requirements.txt
insightface
faiss-cpu
fastapi
uvicorn
opencv-python
numpy
pillow
tqdm
streamlit
requests
scikit-learn
flask
requests
face_recognition
numpy
